Split aliased and flat importing strategies out of EagerImportTag
